+++ planus_back/src/main/java/com/planus/trip/service/TripService.java
@@ -4,6 +4,6 @@ import com.planus.trip.dto.TripInfoResDTO;
 import com.planus.trip.dto.TripResDTO;
 
 public interface TripService {
-    TripResDTO createTrip(long admin, String startDate, int period, int[] areaId);
+    TripResDTO createTrip(long admin, String startDate, long period, int[] areaId);
     TripInfoResDTO findTripInfo(long tripId);
 }
